initialise with redux-thunk
