+++ db.json
@@ -29,11 +29,6 @@
       "content": "Debugging is twice as hard as writing the code in the first place. Therefore, if you write the code as cleverly as possible, you are, by definition, not smart enough to debug it.",
       "id": "98312",
       "votes": 0
-    },
-    {
-      "content": "testing is hard",
-      "votes": 0,
-      "id": "Xo8mnCl"
     }
   ]
 }
