+++ src/index.js
@@ -1,22 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { createStore, combineReducers } from 'redux'
-import { composeWithDevTools } from 'redux-devtools-extension'
 import { Provider } from 'react-redux'
 import App from './App'
-import anecdoteReducer from './reducers/anecdoteReducer'
-import notificationReducer from './reducers/notificationReducer'
-import filterReducer from './reducers/filterReducer'
-
-const reducer = combineReducers({
-  notification: notificationReducer,
-  anecdotes: anecdoteReducer,
-  filter: filterReducer
-})
-
-const store = createStore(
-  reducer,
-  composeWithDevTools())
+import store from './store'
 
 ReactDOM.render(
   <Provider store={store}>
